+++ src/index.js
@@ -1,28 +1,48 @@
 import debounce from './debounce'
 
+// Figures out the event we are using with the bound element
+function figureOutEvent (attrs, listenTo) {
+  const { value = false } = attrs.getNamedItem('debounce-events') || {}
+  const isArr = Array.isArray(listenTo)
+  const toLowerMap = list => list.map(x => x.toLowerCase())
+
+  // If they set an events attribute that overwrites everything
+  if (value) {
+    return toLowerMap(value.split(','))
+  }
+
+  return isArr ? toLowerMap(listenTo) : [listenTo]
+}
+
 export default {
   name: 'debounce',
   install (Vue, { lock, listenTo = 'onkeyup' } = {}) {
     Vue.directive('debounce', {
       bind (el, { value, arg, modifiers }) {
-        if (typeof el[listenTo] === 'undefined') {
-          throw new Error(`Event Listener ${listenTo} does not exist`)
-        }
+        const listener = figureOutEvent(el.attributes, listenTo)
         const fn = debounce(target => {
           value(target.value)
         }, arg)
-        const isUnlocked = (!modifiers.lock && !lock) || modifiers.unlock
 
-        el[listenTo] = ({ keyCode, target }) => {
-          if (keyCode === 13 && isUnlocked) {
+        function handler ({ key, target }) {
+          const isUnlocked = (!modifiers.lock && !lock) || modifiers.unlock
+
+          if (key === 'Enter' && isUnlocked) {
             fn.cancel()
             value(target.value)
           }
 
-          if (keyCode !== 13) {
+          if (key !== 'Enter') {
             fn(target)
           }
         }
+
+        listener.forEach(e => {
+          if (typeof el[e] === 'undefined') {
+            throw new Error(`Event Listener ${e} does not exist`)
+          }
+          el[e] = handler
+        })
       }
     })
   }